Durgod K3x0: Refactoring shared durgod/k310 and durgod/k320 code into durgod/k3x0. Combining chimera keymap with default_toggle_mac_windows keymap
