--- keyboards/durgod/k310/keymaps/default_toggle_mac_windows/rules.mk
@@ -1,1 +1,0 @@
-OPT_DEFS += -DWINLOCK_DISABLED
