--- keyboards/durgod/k310/keymaps/via/readme.md
@@ -1,6 +1,0 @@
-# The default keymap for Durgod Taurus K310
-
-Layer 0 : Standard layout
-
-Layer 1 : Media control
-- Reusing Durgod's Original Media Control for Fn + F1 ~ Fn + F7
