--- keyboards/durgod/k310/keymaps/via/rules.mk
@@ -1,4 +1,0 @@
-OPT_DEFS += -DWINLOCK_DISABLED
-
-VIA_ENABLE = yes
-LTO_ENABLE = yes
