+++ keyboards/durgod/k3x0/k310/rules.mk
@@ -1,0 +1,1 @@
+# see common rules.mk
