+++ keyboards/durgod/k3x0/readme.md
@@ -1,0 +1,13 @@
+# K3X0 = K310, K320, etc.
+
+This K3X0 code is shared between both the ANSI and ISO variants of both the
+K310 full-sized 104/105-key and K320 TKL 87/88-key keyboards.
+
+* Keyboard Maintainers: [dkjer](https://github.com/dkjer) and [tylert](https://github.com/tylert)
+* Hardware Supported: Durgod Taurus K310 or K320 board with STM32F070RBT6
+* Hardware Availability: K310 https://www.durgod.com/page9?product_id=53&_l=en
+                         and K320 https://www.durgod.com/page9?product_id=47&_l=en
+
+Instructions for K310 and K320 building and flashing can be found here:
+* [K310](k310/readme.md)
+* [K320](k320/readme.md)
